requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
Flask==3.0.3
kafka-python==2.0.2
pydantic==2.9.2
PyYAML==6.0.2
python-dotenv==1.0.1
prometheus-client==0.20.0
redis==5.0.8
